Refactor: modal components
