+++ src/providers/ModalProvider.tsx
@@ -1,4 +1,4 @@
-import LoginModal from '../components/LoginModal'
+import LoginModal from '../components/modals/LoginModal'
 
 const ModalProvider = () => {
     return (
